Use ExMachina instead of AccountsFixtures
We use ExMachina instead of AccountsFixtures to modify attributes more
easily. Because a lot of our tests deal with users, our `user_factory`
is a little more complex than usual.

In order for `insert(:user)` to work properly, we need to run it through
the `User.registration_changeset/2`. So our factory uses the changeset
and applies the changes to create a `hashed_password`:

```elixir
%User{}
|> User.registration_changeset(user_attrs)
|> Ecto.Changeset.apply_changes()
```

The only downside is that several tests need the user's password to sign
in, reauthenticate, confirm password, etc. But since we run our
attributes through `User.registration_changeset/2`, the password field
gets removed. To account for that, any test that requires to know the
users password can just pass it into the user (e.g. `insert(:user, password:
"secret password")`). The factory then takes the attributes, merges them
before running them through the registration changeset, and then returns
the applied changes. So our final factory looks like this:

```elixir
def user_factory(attrs) do
  default = %{
    email: sequence(:email, &"user#{&1}@example.com"),
    password: "hello world!"
  }

  user_attrs = merge_attributes(default, attrs)

  %User{}
  |> User.registration_changeset(user_attrs)
  |> Ecto.Changeset.apply_changes()
end
```

We also had to update user registration to run `hash_password/1` outside
of `prepare_changes`. There's an issue open in `phx.gen.auth` that
speaks to this. The problem in the issue is that it's not always desired
to hold a transaction when dealing with hashing passwords. Apparently,
it was put in `prepare_changes` originally as a convenience, so that we
don't do that kind of hashing when validating with LiveView. But we can
simply do the hashing when the changeset is valid. That's what we do in
our case, and that's what Jose reports doing as well. In our case, the
change was driven by a desire to have a clean `user_factory` that we can
simply insert by `insert(:user)`.