Add Auth via phx.gen.auth
We add authentication via `mix phx.gen.auth`.

We remove the header since we know we do not want it hovering of the
application once we register and sign in. We do copy the "Log out" link
to the settings/edit page so we have a way to log out.