Add name, username, avatar_url to users
We add name, username, and avatar_url to users.

We require name and username for account registration, so we update the
form as well as the registration changeset.

We also require uniqueness of the username. We use citext for the
database column and add validation constraints for it in the schema.

We generate the avatar_url based on the user's email to get the avatar
from gravatar, and using ui-avatars as a fallback (which will use the
initials).