Associate posts and users
We add the association between posts and users, where each post must
belong_to a user. We update the post creation logic so that a user_id
must be present, and we update the `Timeline.list_posts/1` and
`Timeline.get_post!` to preload a user since we always want to have the
user information associated to a post when dealing with the Timeline.

We also update the `TimelineTest` to use our factory rather than relying
on the `insert` functions we had created within the file.