Adds Timeline.like_post/2
What?
====

We add a `Timeline.like_post/2` function to handle the liking of a post.
It creates a `%Timeline.Like{}` record with the user and the post, and
it updates the counter cache of `post.likes_count`. That counter cache
is updated via the `Like.changeset/2` which uses Ecto's
`prepare_changes/2` callback. We update the record directly in the DB
with `update_all(query, inc: [likes_count: 1])`.

This is exactly the example that `prepare_changes/2` has in the
documentation -- updating a counter cache.