Render highlighted post
What changed?
=============

We render a highlighted post when a user clicks on it. We use a
`phx-click` attribute approach to doing so. Our `TimelineLive` now has
the concept of a `current_post`.

As we test-drive this feature, we learned about the value of `element/3`
combined with a `render_click/1`. We also talked about the alternative
approach of using `render_click/3` and how using it without an
`element/3` can lead to false positives in our test.