Test navigation to regular (non-live) and live views
We test (and add) navigation from LiveViews to non-liveviews as well as
to other live views.

We consider the navigation from TimelineLive to the user's settings edit
page. And we then consider the navigation from the timeline to the
compose page (another live view).

While doing this, we see how we can use the `{:error, {:redirect, %{to:
path}}}` response from `render_click` to assert which path we've
redirected to or we can use `follow_redirect/3` to follow the
redirection and make an assertion about the path we've landed on. We
also saw how `follow_redirect/3` is smart enough to know whether it
redirected to a regular page or a LiveView, returning `{:ok, conn}` in
the first instance and `{:ok, new_live_view, new_html}` in the second.