Test user can create posts
What changed?
============

We test (and implement) that a user can create a post.

We see the importance of using the `form/3` helper with
`render_submit/1`. If we use `render_submit/3` directly -- targeting the
event and passing the form data -- it's possible to get false positives.

We also see how to test the successful and the error case. One redirects
us to the Timeline, the other we can simply make assertions about the
text that is re-rendered in the ComposeLive