Test new posts creation notifications
What changed?
=============

We test that a user's TimelineLive gets notified of new posts when
another user creates one. We use PubSub as a mechanism to send events,
and we talk about three possible strategies to send the message to the
LiveView process from the test.

In the end, we choose a strategy whereby the test triggers the PubSub
broadcast, but the knowledge of the topic and event format are
centralized in a single module -- thus giving us a single source of
truth.

We centralize the knowledge in the `Timeline` module, but it could have
been centralized equally effectively in a place like
`Wattwitter.PubSub`.