Turn PostComponent into stateful component
We turn the `PostComponent` into a stateful component so that it can
handle the "like" event.

As we do so, we see that the complete logic for a "user liking a post"
lives across two modules: the post component and the timeline live
modules. The post component handles the liking of the post, but the
timeline live handles the updating of the state.

Thus, we see that the "user can like a post" test is difficult to place
in either module. The behavior exists across modules. In fact, it is
process level behavior that spans two modules. And that's why it's
tough to put the test in a single place.