Adds Timeline.list_post_ids/2 and get_posts/2
What changed?
============

This commit adds two functions to `Timeline`:

- `list_post_ids/2` and
- `get_posts/2`

Both functions take in a set of pagination options as the second
argument.

Why do we need them?
----------------------

We introduce them because we'll need them in our work with
`TimelineLive`, `PostComponent`, and `ShowPostComponent`.

Since we're moving the source of truth from the live view to the live
components, we need to fetch the ids of interest (that's where
`Timeline.list_post_ids/2` is useful), and then we'll want to fetch a
list of posts based on those ids -- and that's where
`Timeline.get_posts/2` will be used.