Move user can like a post test
We move the user can like a post test from the `timeline_live_test` to
the `post_component_test`.

After refactoring the source of truth for posts into the live
components, the `PostComponent` can now handle the full "like" event
without having to communicate with the parent live view. Thus, the
module seems responsible for the "liking" behavior.

We note that the test does not change (aside from my intentional change
to logging in the user). That is, testing interactivity with a live
component can only be tested in the same way interactivity is tested in
a live view: we must mount the live view and target the component so
that an action is performed.