Handle post updates in TimelineLive
We add a `handle_info/2` callback to `TimelineLive` to handle post
updates in our `TimelineLive`. Since the `PostComponent` is responsible
for each post, we simply forward the update to the appropriate post via
`send_update/2`.

The test we add in `timeline_live_test.exs` ensures that our
`TimelineLive` live views _aside from the one who made the like_ can
receive the message. We know that the `PostComponent` has a test that
fails when the `TimelineLive` doesn't have a `handle_info/2` for the new
message (and so implicitly tests that timeline live views can handle the
update). But I still like having this test since it defines the contract
other live views must abide by to receive updates when someone likes a
post.