Making tests easier to read
What changed?
============

We update our tests to make them easier to read. We do that by creating
helper functions that encapsulate specific CSS selectors and `element/3`
helpers. The result is that our tests can more easily be read in the
domain of our application without having to concern ourselves with the
specifics of how we target certain elements on the page.

We also point out how helpful it is to separate the 3 steps in each
test: setup (or arrange), action, assertion. We didn't have to rewrite
any tests to fit that since they were already written with that in mind.