Remove pagination from Timeline.get_posts/2
What changed?
============

We remove the pagination from `Timeline.get_posts/2` because it is
unneeded and can lead to subtle bugs.

Since `get_posts/2` takes a list of ids, it is expected that we would
get all the posts for those ids. With pagination defaulting to 10 posts,
it can be confusing (and bug-prone) to only fetch 10 posts when someone
passes more than 10 ids.