Add and test JS InfiniteScroll hook
What changed?
=============

We add and test a JavaScript InfiniteScroll hook.

We use `element/3` with `render_hook/2` and talk about the limitations
of testing hooks. We can only test the Elixir side, but it's not all
bad. Those trade-offs are what allows LiveView tests to be really fast
and robust. If we need to fully exercise the JS, we can use something
like Wallaby or Hound.

We also talk about removing mystery guests in our test setup. We take an
optional `per_page` params so that our test setup does not depend on a
magic 11 posts (10 by default + 1 so we load more posts), but rather,
our test is clear about the number of posts that are rendered per page.