Add test with Floki to ensure new post position
What changed?
============

We add a test with Floki to show how we can assert the position of posts
relative to each other.

To see why that might be helpful, we consider a potential bug if we had
accidentally appended the new post ids to the list of ids (instead of
prepended).

```diff
   def handle_event("show-new-posts", _, socket) do
     {:noreply,
      socket
-     |> update(:post_ids, fn ids -> socket.assigns.new_post_ids ++ ids end)
+     |> update(:post_ids, fn ids -> ids ++ socket.assigns.new_post_ids end)
      |> assign(:new_post_ids, [])
      |> assign(:new_posts_count, 0)}
   end
```

Our existing tests did not enforce that the new posts should be listed
before the older posts. With Floki, we were able to fetch all the posts
and look at their relative positions.