Add test (CSS selectors) to ensure new post in top
What changed?
============

For those interested, even what we did in the test with Floki could have
been done with CSS selectors.

In this commit, we introduce a test that uses the `:first-of-type` CSS
pseudo class to determine which post shows first. We combine that with
the post's body so that our assertion is looking for the first post with
the given text.

Thus, this is another remainder that CSS selectors are very powerful and
we can use them as much as possible before we have to drop down to
Floki.