Test that user can upload an images to client-side
What changed?
============

We start our foray into testing uploads by testing uploads before
submitting the post. That includes the following:

- We test that a user can that a user can upload an image for a post and
  see a preview of that image.
- We test that a user can cancel one of those uploaded images
- Finally, we test that a user gets helpful error messages related to
  those uploads (e.g. "Too many files").

We use the wonderful `live_file_input/2` and `live_img_preview/2`
helpers to accomplish that, and on the testing side, we use
`file_input/3` and `render_upload/2`.

Note that we haven't tested that the images are actually saved. That
will happen in a future commit.