Test Uploads Storage (direct-to-server)
What changed?
=============

We test and implement that we can post a watweet with images. To do
that, we need to make sure we store the images. In this case, we store
those images in our server (direct-to-server).

We use the `consume_uploaded_entries/3` helper to copy the uploads to
their permanent directory and to generate the static paths to store in
our `%Post{}`.

We also update the `PostComponent` and `ShowPostComponent` to render
images with their posts.