Separate photo_urls generation from consuming uploads
What changed?
============

We discover that our users have a bad user experience when posting a
watweet fails -- their uploads disappear!

We capture the bad user experience with a test in our test suite so that
it is treated like a bug.

The bad user experience happens because we're consuming our uploads too
soon. We should only do it if we successfully create a post.

To do that, we create two helper functions: `generate_photo_urls/1`
which uses `uploaded_entries/2` to get the list of completed entries,
and we move the rest of the consuming of uploaded entries into
`save_photos/1`. And now, we only call `save_photos/1` if we've
successfully created a post.