Test that we store files by rendering images
What changed?
============

We want to test that we're actually saving files (`save_photos/1`) in
our `ComposeLive`.

It turns out that's not such an easy thing to do.

We examine three alternatives:

1. Not test it
2. Test it by accessing the file system
3. Test it by requesting the image and expecting a 200

We opt for (3) because even thought option (2) works, we want to keep
our tests from the perspective of the user. Our users can't see our file
system, nor do they care where the files are stored. They just want to
_see_ the images.

Thus, we emulate the behavior of a browser by requesting the image. If
it's a 200, we assume the image will be shown. If it's not a 200, we
assume the user would see a broken image.